% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{dftab}
\alias{dftab}
\title{Convert a data frame into a matrix suitable for barplot input}
\usage{
dftab(x)
}
\arguments{
\item{x}{a data frame with at least one numeric and one factor column and no more than three columns}
}
\value{
a matrix suitable for use with \code{barplot} or \code{nvbarplot}
}
\description{
Convert a data frame into a matrix suitable for barplot input
}
\note{
The data frame must contain at least one factor and one numeric variable, and a total of
  at most three variables. In the case of two factor variables, the first one defines the x-axis
  barplot groups and the 2nd one the within-bar groups.
}
