% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{multibar_matrix2json}
\alias{multibar_matrix2json}
\title{Prepare a barplot matrix for nvd3}
\usage{
multibar_matrix2json(M)
}
\arguments{
\item{M}{matrix}
}
\value{
A custom JSON string suitable for nvd3 multibar plotting
}
\description{
Prepare a barplot matrix for nvd3
}

