% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvmulti.r
\name{nv_bar}
\alias{nv_bar}
\title{nv_bar
Define a bar plot object for use with nvmulti plots.}
\usage{
nv_bar(x, yAxis = 1)
}
\arguments{
\item{x}{a vector of heights or a matrix where each row
represents a vector of heights}

\item{yAxis}{which y-Axis to use}
}
\description{
nv_bar
Define a bar plot object for use with nvmulti plots.
}
\note{
When \code{x} is a matrix, its row names will be used to identify
  bars. However, column names are not yet supported to identify grouped
  sets of bars which instead will be numbered in order.
}
