% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvmulti.r
\name{nv_coords}
\alias{nv_coords}
\title{nv_coords
Define a scatter or line plot object for use with nvmulti plots.}
\usage{
nv_coords(x, y, key = "data", yAxis = 1, type = c("scatter", "line"))
}
\arguments{
\item{x}{A vector of x-axis coordinates}

\item{y}{A vector of y-axis coordinates}

\item{key}{a character label for the scatter plot}

\item{yAxis}{which y-Axis to use}

\item{type}{the type of plot, either "scatter" or "line"}
}
\description{
nv_coords
Define a scatter or line plot object for use with nvmulti plots.
}
