% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvarea.r
\name{nvareaplot}
\alias{nvareaplot}
\title{nv stacked area plot}
\usage{
nvareaplot(x, nxticks = 10, xticklabels = NULL, ytickformat = ",.1f",
  interpolate = c("step", "linear", "basis", "step-before", "step-after",
  "bundle", "cardinal", "monotone"), controls = FALSE, xlim, ylim, tooltip,
  options)
}
\arguments{
\item{x}{a data frame with at least two columns whose first column contains
numeric x-axis values and remaining columns contain stacked area y-axis values.}

\item{nxticks}{number of x tick marks, set to 0 for none.}

\item{xticklabels}{either a character-valued function of a single numeric value
that converts x-axis values to text, or a vector of labels as long as \code{nxticks}.}

\item{ytickformat}{format numeric tick labels using a d3.format string (d3.js)}

\item{interpolate}{d3.js plot interpolation option}

\item{controls}{(logical), if \code{TRUE} then show optional interactive plot format controls}

\item{xlim}{an optional numeric vector of two values of lower and upper x-axis limits}

\item{ylim}{an optional numeric vector of two values of lower and upper y-axis limits}

\item{tooltip}{an optional character vector with as many entries as \code{x} has rows, the associated entry is displayed when the user's mouse hovers over the corresponding x-xaxis value (assumes that \code{x} has distinct x-axis values). The character values may be formatted with HTML formatting.}

\item{options}{optional additional JavaScript options passed directly to nvd3.js.}
}
\value{
An htmlwidget object that is displayed using the object's show or print method.
(If you don't see your widget plot, try printing it with the \code{print} function.)
}
\description{
nv stacked area plot
}
\examples{
data("stackedArea", package="nvd3")
head(stackedArea)  # note POSIX date format in 1st column
nvareaplot(stackedArea)

# Now plot with a better x-axis format:
nvareaplot(stackedArea, xticklabels=function(x)
             format(as.POSIXct(x, origin="1970-1-1"), "\%Y-\%m-\%d"))

}

