% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{nvd3-package}
\alias{nvd3-package}
\alias{nvd3}
\title{Basic interactive plots using nvd3.js and htmlwidgets.}
\description{
Basic interactive plots using nvd3.js and htmlwidgets.
}
\examples{
\dontrun{
library("shiny")

# See also help for nvd3.js
}

}
\references{
\url{http://nvd3.org}
}
