% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{nvd3}
\alias{nvd3}
\title{Render generic nvd3 plots}
\usage{
nvd3(program, height = NULL, width = NULL, ...)
}
\arguments{
\item{program}{JavaScript program to draw plot}

\item{height}{Widget height}

\item{width}{Widget width}

\item{...}{Optional additional parameters passed to the rendering code.}
}
\value{
An HTML widget object
}
\description{
Render generic nvd3 plots
}

