% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{stackedArea}
\alias{stackedArea}
\title{Stacked Area Plot Example Data}
\format{
A data frame.
}
\source{
\url{http://nvd3.org/examples/stackedArea.html}.
}
\usage{
data(stackedArea)
}
\description{
Stacked Area Plot Example Data
}
\keyword{datasets}
