% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvmulti.r
\name{nv_area}
\alias{nv_area}
\title{nv_area}
\usage{
nv_area(x, yAxis = 1)
}
\arguments{
\item{x}{a matrix with at least 2 columns whose first column contains x-coordinate values
and remaining columns series y-coordinate values}

\item{yAxis}{which y-Axis to use}
}
\description{
nv_area
}

