% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvmulti.r
\name{nvmulti}
\alias{nvmulti}
\title{nvmulti
Overylay scatter, line, bar, and stacked area plots on one or two vertical axes.}
\usage{
nvmulti(..., tooltip = FALSE, guideline = TRUE, interpolate = c("linear",
  "step", "basis", "step-before", "step-after", "bundle", "cardinal",
  "monotone"), nxticks = 10, col, xlim, ylim1, ylim2, xticklabels,
  options = "")
}
\arguments{
\item{...}{\code{nv_coords}, \code{nv_bar}, and/or \code{nv_area} objects to plot.}

\item{tooltip}{set to TRUE to display mouse hover point summary tooltip}

\item{guideline}{set to TRUE to display mouse hover summary of all series}

\item{interpolate}{d3.js plot interpolation option}

\item{nxticks}{number of x tick marks, set to 0 for none.}

\item{col}{optional colors, one for each series}

\item{xlim}{an optional numeric vector of two values of lower and upper x-axis limits}

\item{ylim1}{an optional numeric vector of two values of lower and upper left y-axis limits}

\item{ylim2}{an optional numeric vector of two values of lower and upper right y-axis limits}

\item{xticklabels}{optional, either a character-valued function of a single numeric value
that converts x-axis values to text, or a vector of labels as long as \code{nxticks}.}

\item{options}{advanced JavaScript options}
}
\description{
nvmulti
Overylay scatter, line, bar, and stacked area plots on one or two vertical axes.
}

